requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.0.1

# Vector database
pinecone-client>=2.2.4

# Document processing
unstructured>=0.10.30
pymupdf>=1.23.7

# LLM and embeddings
openai>=1.3.0

# Utilities
loguru>=0.7.0
pydantic>=2.4.0
python-dotenv>=1.0.0

# Web application
streamlit>=1.28.0